fix plot_lines
